+++ src/main/java/com/unstar/backend/domain/entity/Comment.java
@@ -15,7 +15,7 @@ public class Comment extends CommonVO implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false)
-    private Long id;                              // 고유번호
+    private Integer id;                              // 고유번호
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "board_id")
