+++ src/main/java/com/unstar/backend/dto/response/CommentResponseDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CommentResponseDTO {
-    private Long id; //고유번호
+    private Integer id; //고유번호
     private String userName; //글쓴이
     private String content; //댓글 내용
 }
